requirements.txt
flask
requests
beautifulsoup4
groq
markdown
python-dotenv
